+++ projectroles/templatetags/projectroles_common_tags.py
@@ -11,11 +11,14 @@ from django.template.loader import get_template
 from django.urls import reverse
 
 import projectroles
-from projectroles.models import Project
+from projectroles.models import Project, RemoteProject, SODAR_CONSTANTS
 from projectroles.plugins import get_backend_api
 
 site = import_module(settings.SITE_PACKAGE)
 User = get_user_model()
+
+SITE_MODE_SOURCE = SODAR_CONSTANTS['SITE_MODE_SOURCE']
+SITE_MODE_TARGET = SODAR_CONSTANTS['SITE_MODE_TARGET']
 
 register = template.Library()
 
@@ -95,11 +98,35 @@ def get_user_html(user):
 
 
 @register.simple_tag
-def get_project_link(project, full_title=False):
+def get_project_link(project, full_title=False, request=None):
     """Return link to project with a simple or full title"""
-    return '<a href="{}">{}</a>'.format(
+    remote_icon = ''
+
+    if request:
+        remote_icon = get_remote_icon(project, request)
+
+    return '<a href="{}">{}</a> {}'.format(
         reverse('projectroles:detail', kwargs={'project': project.sodar_uuid}),
-        project.get_full_title() if full_title else project.title)
+        project.get_full_title() if full_title else project.title,
+        remote_icon)
+
+
+@register.simple_tag
+def get_remote_icon(project, request):
+    """Get remote project icon HTML"""
+    if (settings.PROJECTROLES_SITE_MODE == SITE_MODE_TARGET and
+            request.user.is_superuser):
+        try:
+            remote_project = RemoteProject.objects.get(project=project)
+            return '<i class="fa fa-lock text-muted mx-1 ' \
+                   'sodar-pr-remote-project-icon" title="{}" ' \
+                   'data-toggle="tooltip" data-placement="top"></i>'.format(
+                    'Remote project from ' + remote_project.site.name)
+
+        except RemoteProject.DoesNotExist:
+            pass
+
+    return ''
 
 
 @register.simple_tag